refactor: Change Room to Place
- Room을 Place로 수정하였습니다.
- 단순히 방 하나만이 아니기 때문에 수정.
