+++ BE/src/main/java/codesquad/team17/gnb/place/Place.java
@@ -1,11 +1,11 @@
-package codesquad.team17.gnb.room;
+package codesquad.team17.gnb.place;
 
-public class Room {
+public class Place {
     private Long id;
     private String name;
     private int maximumNumberOfPeople;
 
-    public Room(Long id, String name, int maximumNumberOfPeople) {
+    public Place(Long id, String name, int maximumNumberOfPeople) {
         this.id = id;
         this.name = name;
         this.maximumNumberOfPeople = maximumNumberOfPeople;
