+++ BE/src/main/java/codesquad/team17/gnb/place/PlaceRepository.java
@@ -1,0 +1,9 @@
+package codesquad.team17.gnb.place;
+
+import java.util.Optional;
+
+public interface PlaceRepository {
+
+    Optional<Place> findById(Long id);
+
+}
