--- BE/src/main/java/codesquad/team17/gnb/room/RoomRepository.java
@@ -1,9 +1,0 @@
-package codesquad.team17.gnb.room;
-
-import java.util.Optional;
-
-public interface RoomRepository {
-
-    Optional<Room> findById(Long id);
-
-}
