+++ BE/src/test/resources/data.sql
@@ -1,2 +1,2 @@
-INSERT INTO room (room_id, room_name, maximum_number_of_people) VALUES (1, '서울호스텔', 3);
-INSERT INTO room (room_id, room_name, maximum_number_of_people) VALUES (2, '코드스쿼드', 4);
+INSERT INTO place (place_id, place_name, maximum_number_of_people) VALUES (1, '서울호스텔', 3);
+INSERT INTO place (place_id, place_name, maximum_number_of_people) VALUES (2, '코드스쿼드', 4);
